Base class for Option classes
- Added a base class for the Option classes.
