+++ app/Http/Controllers/RequestController.php
@@ -115,9 +115,9 @@ class RequestController extends Controller
     public function optionsAccessLog(Request $request)
     {
         $get = Get::init()->
-            setDescription('route-descriptions.request_GET_access-log')->
             setParameters('api.request-access-log.parameters.collection')->
             setPagination(true)->
+            setDescription('route-descriptions.request_GET_access-log')->
             option();
 
         return $this->optionsResponse($get, 200);
@@ -136,8 +136,8 @@ class RequestController extends Controller
             option();
 
         $post = Post::init()->
-            setDescription('route-descriptions.request_POST')->
             setFields('api.request-error-log.fields')->
+            setDescription('route-descriptions.request_POST')->
             option();
 
         return $this->optionsResponse(
