+++ app/Option/Patch.php
@@ -13,27 +13,12 @@ use Illuminate\Support\Facades\Config;
  * @copyright G3D Development Limited 2018-2019
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class Patch
+class Patch extends Option
 {
-    /**
-     * @var Patch
-     */
-    static private $instance;
-
-    /**
-     * @var boolean
-     */
-    static private $authentication;
-
     /**
      * @var array
      */
     static private $conditional_fields;
-
-    /**
-     * @var string
-     */
-    static private $description;
 
     /**
      * @var array
@@ -47,28 +32,18 @@ class Patch
 
     static private function reset()
     {
-        self::$authentication = false;
+        self::resetBase();
+
         self::$conditional_fields = [];
-        self::$description = null;
         self::$fields = [];
         self::$localised_fields = [];
     }
 
     static public function init(): Patch
     {
-        if (self::$instance === null) {
-            self::$instance = new Patch();
-            self::$instance->reset();
-        }
 
-        return self::$instance;
-    }
-
-    static public function setAuthenticationRequired(
-        bool $status = false
-    ): Patch
-    {
-        self::$authentication = $status;
+        self::$instance = new Patch();
+        self::$instance->reset();
 
         return self::$instance;
     }
@@ -82,15 +57,6 @@ class Patch
         return self::$instance;
     }
 
-    static public function setDescription(
-        string $localisation_path
-    ): Patch
-    {
-        self::$description = trans($localisation_path);
-
-        return self::$instance;
-    }
-
     static public function setFields(
         string $config_path
     ): Patch
@@ -99,7 +65,7 @@ class Patch
         return self::$instance;
     }
 
-    static private function buildFields()
+    static protected function build()
     {
         self::$localised_fields = [];
 
@@ -119,7 +85,7 @@ class Patch
 
     static public function option(): array
     {
-        self::buildFields();
+        self::build();
 
         return [
             'PATCH' => [
